# openapi/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: oojuniin@outlook.com
    name: Junior Lima
  description: "Contrato REST do budget-service. Este documento é a fonte da verdade\
    \ para qualquer client (frontend, agentes de IA) que consuma esta API — não assuma\
    \ comportamento não descrito aqui."
  title: Budget Service API
  version: 0.0.1
servers:
- description: Relativo ao host onde a API estiver publicada
  url: /
security:
- bearer-jwt: []
paths:
  /api/v1/budget-rules/fifty-thirty-twenty:
    get:
      operationId: fiftyThirtyTwenty
      parameters:
      - in: query
        name: month
        required: true
        schema:
          type: integer
          format: int32
      - in: query
        name: year
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FiftyThirtyTwentyDTO"
          description: OK
      tags:
      - budget-rule-controller
  /api/v1/budget-rules/monthly-summary:
    get:
      operationId: monthlySummary
      parameters:
      - in: query
        name: month
        required: true
        schema:
          type: integer
          format: int32
      - in: query
        name: year
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MonthlySummaryDTO"
          description: OK
      tags:
      - budget-rule-controller
  /api/v1/budget-rules/yearly-summary:
    get:
      operationId: yearlySummary
      parameters:
      - in: query
        name: year
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/YearlySummaryDTO"
          description: OK
      tags:
      - budget-rule-controller
...
  /api/v1/revenue-types/{id}:
    delete:
      operationId: delete_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - revenue-type-controller
    get:
      operationId: findById_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueTypeDTO"
          description: OK
      tags:
      - revenue-type-controller
    put:
      operationId: update_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueTypeDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueTypeDTO"
          description: OK
      tags:
      - revenue-type-controller
  /api/v1/revenue-types/{id}/history:
    get:
      operationId: history_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueTypeRevisionDTO"
          description: OK
      tags:
      - revenue-type-controller
  /api/v1/revenues:
    get:
      operationId: search_1
      parameters:
      - in: query
        name: month
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: year
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: typeId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageRevenueDTO"
          description: OK
      tags:
      - revenue-controller
    post:
      operationId: save_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueDTO"
          description: Created
      tags:
      - revenue-controller
  /api/v1/revenues/batch:
    post:
      operationId: saveAll_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueBatchRequestDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDTO"
          description: Created
      tags:
      - revenue-controller
  /api/v1/revenues/batch/annual:
    post:
      operationId: saveAnnual_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueAnnualBatchRequestDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDTO"
          description: Created
      tags:
      - revenue-controller
  /api/v1/revenues/by-type:
    get:
      operationId: totalByType_1
      parameters:
      - in: query
        name: year
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypeTotalDTO"
          description: OK
      tags:
      - revenue-controller
  /api/v1/revenues/total:
    get:
      operationId: total_1
      parameters:
      - in: query
        name: month
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: year
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: typeId
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TotalDTO"
          description: OK
      tags:
      - revenue-controller
  /api/v1/revenues/{id}:
    delete:
      operationId: delete_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - revenue-controller
    get:
      operationId: findById_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueDTO"
          description: OK
      tags:
      - revenue-controller
    put:
      operationId: update_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueDTO"
          description: OK
      tags:
      - revenue-controller
  /api/v1/revenues/{id}/history:
    get:
      operationId: history_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueRevisionDTO"
          description: OK
      tags:
      - revenue-controller
...
  /api/v1/spending-types/{id}:
    delete:
      operationId: delete_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - spending-type-controller
    get:
      operationId: findById_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingTypeDTO"
          description: OK
      tags:
      - spending-type-controller
    put:
      operationId: update_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendingTypeDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingTypeDTO"
          description: OK
      tags:
      - spending-type-controller
  /api/v1/spending-types/{id}/history:
    get:
      operationId: history_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingTypeRevisionDTO"
          description: OK
      tags:
      - spending-type-controller
...
  /api/v1/spendings/batch:
    post:
      operationId: saveAll
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendingBatchRequestDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingDTO"
          description: Created
      tags:
      - spending-controller
  /api/v1/spendings/batch/annual:
    post:
      operationId: saveAnnual
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendingAnnualBatchRequestDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingDTO"
          description: Created
      tags:
      - spending-controller
  /api/v1/spendings/by-type:
    get:
      operationId: totalByType
      parameters:
      - in: query
        name: year
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypeTotalDTO"
          description: OK
      tags:
      - spending-controller
...
  /api/v1/spendings/{id}:
    delete:
      operationId: delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - spending-controller
    get:
      operationId: findById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingDTO"
          description: OK
      tags:
      - spending-controller
    put:
      operationId: update
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendingDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingDTO"
          description: OK
      tags:
      - spending-controller
  /api/v1/spendings/{id}/history:
    get:
      operationId: history
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingRevisionDTO"
          description: OK
      tags:
      - spending-controller
components:
  schemas:
    BudgetBucketDTO:
      type: object
      properties:
        actual:
          type: number
        difference:
          type: number
        target:
          type: number
    FiftyThirtyTwentyDTO:
      type: object
      properties:
        essential:
          $ref: "#/components/schemas/BudgetBucketDTO"
        personal:
          $ref: "#/components/schemas/BudgetBucketDTO"
        savings:
          $ref: "#/components/schemas/BudgetBucketDTO"
        totalRevenue:
          type: number
    MonthlySummaryDTO:
      type: object
      properties:
        projectedBalance:
          type: number
        totalPaid:
          type: number
        totalPending:
          type: number
        totalRevenue:
          type: number
        totalSpending:
          type: number
    PageRevenueDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDTO"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    RevenueAnnualBatchRequestDTO:
      type: object
      properties:
        months:
          type: array
          items:
            type: integer
            format: int32
            maximum: 12
            minimum: 1
          uniqueItems: true
        revenues:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDTO"
          maxItems: 500
          minItems: 0
      required:
      - revenues
    RevenueBatchRequestDTO:
      type: object
      properties:
        revenues:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDTO"
          maxItems: 500
          minItems: 0
      required:
      - revenues
    RevenueDTO:
      type: object
      properties:
        date:
          type: string
          format: date
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        typeId:
          type: string
          format: uuid
        typeName:
          type: string
        value:
          type: number
      required:
      - date
      - typeId
      - value
    RevenueRevisionDTO:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        date:
          type: string
          format: date
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
        typeId:
          type: string
          format: uuid
        value:
          type: number
    RevenueTypeDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        includeInMonthlyTotals:
          type: boolean
        includeInTotals:
          type: boolean
        name:
          type: string
          minLength: 1
      required:
      - name
    RevenueTypeRevisionDTO:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        id:
          type: string
          format: uuid
        includeInMonthlyTotals:
          type: boolean
        includeInTotals:
          type: boolean
        name:
          type: string
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    SpendingAnnualBatchRequestDTO:
      type: object
      properties:
        months:
          type: array
          items:
            type: integer
            format: int32
            maximum: 12
            minimum: 1
          uniqueItems: true
        spendings:
          type: array
          items:
            $ref: "#/components/schemas/SpendingDTO"
          maxItems: 500
          minItems: 0
      required:
      - spendings
    SpendingBatchRequestDTO:
      type: object
      properties:
        spendings:
          type: array
          items:
            $ref: "#/components/schemas/SpendingDTO"
          maxItems: 500
          minItems: 0
      required:
      - spendings
    SpendingDTO:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
          maxLength: 250
          minLength: 3
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        typeId:
          type: string
          format: uuid
        typeName:
          type: string
        value:
          type: number
        wasPaid:
          type: boolean
      required:
      - date
      - typeId
      - value
      - wasPaid
    SpendingRevisionDTO:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
        typeId:
          type: string
          format: uuid
        value:
          type: number
        wasPaid:
          type: boolean
    SpendingTypeDTO:
      type: object
      properties:
        category:
          type: string
          enum:
          - ESSENTIAL
          - PERSONAL
          - SAVINGS
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
      required:
      - category
      - name
    SpendingTypeRevisionDTO:
      type: object
      properties:
        category:
          type: string
          enum:
          - ESSENTIAL
          - PERSONAL
          - SAVINGS
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
    TotalDTO:
      type: object
      properties:
        total:
          type: number
    TypeTotalDTO:
      type: object
      properties:
        total:
          type: number
        typeId:
          type: string
          format: uuid
        typeName:
          type: string
    YearlySummaryDTO:
      type: object
      properties:
        balance:
          type: number
        totalRevenue:
          type: number
        totalSpending:
          type: number
  securitySchemes:
    bearer-jwt:
      bearerFormat: JWT
      name: bearer-jwt
      scheme: bearer
      type: http